Module Jekyll_format.E

These are error strings returned by the parser. They are primarily used by the test cases to pattern match on failures and are not for general use.

val yaml_no_start : string

val yaml_no_end : string

val yaml_field_parse : string -> string