Module Jekyll_liquid

Jekyll Liquid templates parsing module

Liquid templates are a library for flexible web apps. The original library is in Ruby, and this module is currently just a partial implementation that handles certain tags such as code highlighting.

module Tags : sig ... end

val highlight_exn : ?⁠f:(Tags.highlight option -> Astring.String.sub -> Astring.String.sub) -> Astring.String.sub -> Astring.String.sub

highlight body parses the body for Jekyll ` highlight tags and applies f to them and substitutes the result into the tag body. The default f is highlight_markdown_code that transforms them into vanilla Markdown with no special highlighting except for a code segment.

val highlight_markdown_code : Tags.highlight option -> Astring.String.sub -> Astring.String.sub

highlight_markdown_code s will wrap the code in s in a Markdown code segment. This can be parsed to highlight_exn as a sensible default.